src/backend/requirements.txt
Flask==0.12
boto3==1.16.50
PyJWT==1.4.0
